refactor: move deprecated folder
