+++ apps/web/deprecated/resource.tsx
