+++ apps/web/pages/index.tsx
@@ -1,102 +1,53 @@
-import { useActor } from '@xstate/react'
+import { useActor, useService } from '@xstate/react'
+import { Button, Modal, Table } from 'antd'
 import React, { useContext } from 'react'
-import { BaseNodeType, NodeA } from '@codelab/shared/interface/node'
-import {
-  ContextNode,
-  EventNameNode,
-  EventNode,
-  StateNameNode,
-} from '@codelab/state/node'
-import {
-  FormNode,
-  Layout,
-  MachineContext,
-  Modal,
-  ModalButton,
-  Table,
-} from '@codelab/ui/component'
+import { collectionToTable } from '@codelab/shared/factory'
+import { EventNameApp } from '@codelab/state/app'
+import { FormGraph, MachineContext } from '@codelab/ui/component'
 
-const Index = () => {
+const GraphPage = () => {
   const { actors } = useContext(MachineContext)
-  const [nodeState, nodeSend] = useActor<ContextNode, EventNode>(actors.node)
+  const [appState, appSend] = useService(actors.app)
+  const [graphState] = useActor(actors.graph)
 
-  // const handlecancel = () => {
-  //   actors.modal.send({ type: EventNameModal.CLOSE })
+  // const createVertex = () =>
+  //   mutate<CreateGraphMutation, CreateGraphMutationVariables>(
+  //     getApolloClient(),
+  //     {
+  //       mutation: CreateGraphDocument,
+  //       variables: {
+  //         input: {
+  //           label: 'My Graph',
+  //         },
+  //       },
+  //       context: {
+  //         clientName: 'hasura',
+  //       },
+  //     },
+  //   )
 
-  //   nodeState.value === StateNameNode.EDITING
-  //     ? nodeSend({ type: EventNameNode.NODE_EDIT_CANCEL })
-  //     : noop()
-  // }
-
-  const handlesubmit = (values: any) => {
-    return nodeState.value === StateNameNode.EDITING
-      ? nodeSend({
-          type: EventNameNode.NODE_EDIT_SUBMIT,
-          payload: values,
-        })
-      : nodeSend({
-          type: EventNameNode.NODE_CREATE,
-          payload: values,
-        })
+  const buttonProps = {
+    onClick: () => appSend(EventNameApp.START_CREATE),
   }
-  const initialvalues =
-    nodeState.value === StateNameNode.EDITING
-      ? {
-          nodeType: BaseNodeType.React,
-          ...nodeState.context.editedNode,
-        }
-      : {
-          nodeType: BaseNodeType.React,
-          parent: null,
-        }
-
-  const handleedit = (nodeId: NodeA['id']) =>
-    nodeSend({ type: EventNameNode.NODE_EDIT, payload: nodeId })
-
-  const handledelete = (nodeId: NodeA['id']) =>
-    nodeSend({ type: EventNameNode.NODE_DELETE, payload: nodeId })
 
   const modalProps = {
-    // handlecancel,
-    actor: actors.modal,
-  }
-
-  const formNodeProps = {
-    actor: actors.modal,
-    handlesubmit,
-    initialvalues,
-  }
-
-  const tableProps = {
-    data: nodeState.context.nodes.map((node: any) => ({
-      ...node,
-      key: node.id,
-    })),
-    handleedit,
-    handledelete,
-    selectnode: () => null,
+    title: 'Create Graph',
+    visible: actors.modal.state.context.visible,
+    onOk: () => appSend(EventNameApp.CANCEL),
+    onCancel: () => appSend(EventNameApp.CANCEL),
   }
 
   return (
     <>
-      <Layout
-        actor={actors.layout}
-        content={
-          <>
-            {/* <ReactJson data={actors.modal.state} /> */}
-            <ModalButton actor={actors.modal} />
-            <Modal {...modalProps}>
-              <FormNode {...formNodeProps} />
-            </Modal>
-            <Table {...tableProps} />
-          </>
-        }
-        sidebar={<></>}
-        header={<></>}
-        footer={<></>}
-      />
+      {/* <ReactJson data={graphState.value} /> */}
+      {/* <ReactJson data={collectionToTable(graphState.context.list)} /> */}
+      <Button {...buttonProps}>+ Create New</Button>
+      <Modal {...modalProps}>
+        <FormGraph />
+      </Modal>
+      <Table {...collectionToTable(graphState.context.list)} />
     </>
   )
 }
 
-export default Index
+export default GraphPage
