+++ apps/web/tsconfig.json
@@ -13,7 +13,7 @@
     "isolatedModules": true
   },
   "include": ["**/*.ts", "**/*.tsx", "**/*.graphql", "next-env.d.ts"],
-  "exclude": ["node_modules"],
+  "exclude": ["node_modules", "**/deprecated/**/*"],
   "references": [
     {
       "path": "./.storybook/tsconfig.json"
